fix: update create map
on the create map, you can click on the map to add a checkpoint, which is updates the state of the checkpoints array. Next step is to handle clicks better to prevent unwanted placements of checkpoints, then to prompt the user if they want to confirm and save their course once they are done.